Factor out Argument class
