+++ index.js
@@ -1,11 +1,9 @@
-/**
- * Module dependencies.
- */
-
 const EventEmitter = require('events').EventEmitter;
 const childProcess = require('child_process');
 const path = require('path');
 const fs = require('fs');
+
+const { Argument } = require('./lib/argument');
 
 // @ts-check
 
@@ -377,81 +375,6 @@ class Help {
       return ((i > 0) ? indentString : '') + line.trimRight();
     }).join('\n');
   }
-}
-
-class Argument {
-  /**
-   * Initialize a new command argument with the given name and description.
-   * The default is that the argument is required, and you can explicitly
-   * indicate this with <> around the name. Put [] around the name for an optional argument.
-   *
-   * @param {string} name
-   * @param {string} [description]
-   */
-
-  constructor(name, description) {
-    this.description = description || '';
-    this.variadic = false;
-    this.parseArg = undefined;
-    this.defaultValue = undefined;
-    this.defaultValueDescription = undefined;
-
-    switch (name[0]) {
-      case '<': // e.g. <required>
-        this.required = true;
-        this._name = name.slice(1, -1);
-        break;
-      case '[': // e.g. [optional]
-        this.required = false;
-        this._name = name.slice(1, -1);
-        break;
-      default:
-        this.required = true;
-        this._name = name;
-        break;
-    }
-
-    if (this._name.length > 3 && this._name.slice(-3) === '...') {
-      this.variadic = true;
-      this._name = this._name.slice(0, -3);
-    }
-  }
-
-  /**
-   * Return argument name.
-   *
-   * @return {string}
-   */
-
-  name() {
-    return this._name;
-  };
-
-  /**
-   * Set the default value, and optionally supply the description to be displayed in the help.
-   *
-   * @param {any} value
-   * @param {string} [description]
-   * @return {Argument}
-   */
-
-  default(value, description) {
-    this.defaultValue = value;
-    this.defaultValueDescription = description;
-    return this;
-  };
-
-  /**
-   * Set the custom handler for processing CLI command arguments into argument values.
-   *
-   * @param {Function} [fn]
-   * @return {Argument}
-   */
-
-  argParser(fn) {
-    this.parseArg = fn;
-    return this;
-  };
 }
 
 class Option {
